work orders display if filtered or not show correct error
