+++ src/components/Property/WorkOrdersHistoryFilter/Index.tsx
@@ -27,7 +27,7 @@ const WorkOrdersHistoryFilter = ({
 
   return (
     <div className="trade-picker-container">
-      <label htmlFor="trade-picker" className="lbh-heading-h2">
+      <label htmlFor="trade-picker" className="lbh-heading-h4">
         Filter by trade:
       </label>
       {trades !== null && (
